move boxed stream into main consumer api
